requirements.txt
pillow 

# PDF extraction
pdfplumber
pytesseract
pdf2image
PyMuPDF

# Boto3 for AWS S3 access
boto3

# Load Env
python-dotenv

# LangChain core
langchain
langchain-text-splitters
langchain-community
langchain-core
langchain-aws

# Vector database
faiss-cpu

# Data models & validation
pydantic

# Utilities
numpy

# Better PDF table extraction (financial bids)
camelot-py[cv]

# Logging / debugging
loguru

# UI dashboard (if you plan to demo)
streamlit

# Backend
fastapi
uvicorn
python-multipart

sentence_transformers